requirements.txt
streamlit
openai
crewAI
langchain
youtube-transcript-api
fpdf2
langchain-openai
langchain-community
pdfplumber
weasyprint
langchain-google-genai 
google-generativeai
